requirements.txt
autopep8==1.6.0
boto3==1.22.4
botocore==1.25.4
jmespath==1.0.0
Pillow==9.1.0
pycodestyle==2.8.0
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
toml==0.10.2
urllib3==1.26.9